requirements.txt
fastapi
uvicorn
streamlit
google-generativeai
groq
sqlalchemy
python-dotenv
requests
pandas
pydantic
fpdf2
email-validator
